Trying to fix the problems
I think the problem with the death effect is that the photonnetwork isn't recognizing the other player as a Player, so they aren't in the rpc list. I've asked for help, and if i'm right, if i fix the player manager thing, i should fix the death effect problem as well.