Found this bug with my game
Trying to fix it. Here's the bug description. Now, players can't die from other clients because of some sort of photon view problem. The second bug is that when you die from fall damage, and respawn and try to leave, it doesn't work.

SOME THINGS TO TRY TO FIX IT:

*See if one client can view another client dying from fall damage.
        If so, procceed with new info.