I fixed the issue specified in the last commit
Here's the issue I fixed:

When I'm in a room as a player, and I leave the room, the code tells the game to move to scene 0. However, when player moves to scene 0, it opens up the LogIn canvas again, even if player is logged in. Need to implement system to check if player is logged in. If logged in, disable account canvas and enable menu canvas