I think I fixed the account system bug (read desc)
Here was the actual bug:

When I login, open a match, then leave the match, when I loaded the main menu, it gave me an error saying that something wasn't being referenced properly, and also the PhotonNetwork.Nickname wasn't saving, and nothing was saving. It was a pretty confusing bug. I think I have a fix for it.

I commented out the "auto-login" code which is supposed to run when the player joins a match and leaves. Since I commented it out, the auto login isn't working, but at least I'm not getting errors. The next step in fixing this bug is to make a new "auto-login" system.