I now have added correct player color syncing ingame!
There's a bug I've found though

When I log in, enter a match, leave the match, sign out, log in with a different account, the main menu doesn't open!

I think its because when you leave the match, you disconnect? So, when you enter the main menu, I think I have to add some code that reconnects when you get back to the main menu.